requirements.txt
Flask==1.0.2
google-cloud-speech==0.36.3
google-cloud-storage==1.13.2
gunicorn==19.9.0
requests
BeautifulSoup4
pyaudio